+++ app/controllers/collection_objects_controller.rb
@@ -467,7 +467,7 @@ class CollectionObjectsController < ApplicationController
       collecting_event_ids: [],
       collector_ids: [], #
       determiner_id: [],
-      geographic_area_ids: [],
+      geographic_area_id: [],
       is_type: [],
       keyword_id_and: [],
       keyword_id_or: [],
@@ -545,7 +545,7 @@ class CollectionObjectsController < ApplicationController
       collecting_event_ids: [],
       collector_ids: [],
       determiner_id: [],
-      geographic_area_ids: [], # CE filter
+      geographic_area_id: [], # CE filter
       is_type: [],
       keyword_id_and: [],
       keyword_id_or: [],