+++ app/models/taxon_name_classification/iczn/available/invalid/synonymy_of_type_genus_before_1961.rb
@@ -1,9 +1,9 @@
-class TaxonNameClassification::Iczn::Available::Invalid::HomonymyOfTypeGenus < TaxonNameClassification::Iczn::Available::Invalid
+class TaxonNameClassification::Iczn::Available::Invalid::SynonymyOfTypeGenusBefore1961 < TaxonNameClassification::Iczn::Available::Invalid
 
-  NOMEN_URI='http://purl.obolibrary.org/obo/NOMEN_0000119'.freeze
+  NOMEN_URI='http://purl.obolibrary.org/obo/NOMEN_0001061'.freeze
 
   def classification_label
-    'invalid due to homonymy of type genus'
+    'invalid due to synonymy of type genus before 1961'
   end
 
   def self.applicable_ranks
@@ -14,7 +14,7 @@ class TaxonNameClassification::Iczn::Available::Invalid::HomonymyOfTypeGenus < T
     self.parent.disjoint_taxon_name_classes +
         self.collect_to_s(TaxonNameClassification::Iczn::Available::Invalid,
                           TaxonNameClassification::Iczn::Available::Invalid::Homonym,
-                          TaxonNameClassification::Iczn::Available::Invalid::SynonymyOfTypeGenusBefore1961,
+                          TaxonNameClassification::Iczn::Available::Invalid::HomonymyOfTypeGenus,
                           TaxonNameClassification::Iczn::Available::Invalid::SuppressionOfTypeGenus)
   end